teamPage 이름 변경 & teamMaker 파일 분리
